Fix qConstexprNextPowerOfTwo(qint64) to return quint64
It doesn't make sense to lose precision, and that's a fatal warning
when used on a 32-bit platform:
implicit conversion loses integer precision: 'quint64' (aka 'unsigned long long')
  to 'quint32' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]

Also fix coding style.

Amends 7ef382649754c261ca9eb99dd50b67050e748efb

+++ src/corelib/kernel/qmath.h
@@ -265,14 +265,16 @@ constexpr inline quint64 qConstexprNextPowerOfTwo(quint64 v) {
     return v;
 }
 
-constexpr inline quint32 qConstexprNextPowerOfTwo(qint32 v) {
+constexpr inline quint32 qConstexprNextPowerOfTwo(qint32 v)
+{
     return qConstexprNextPowerOfTwo(quint32(v));
 }
 
-constexpr inline quint32 qConstexprNextPowerOfTwo(qint64 v) {
+constexpr inline quint64 qConstexprNextPowerOfTwo(qint64 v)
+{
     return qConstexprNextPowerOfTwo(quint64(v));
 }
-}
+} // namespace QtPrivate
 
 constexpr inline quint32 qNextPowerOfTwo(quint32 v)
 {
